Fixed focus of input

+++ resources/js/components/ShoppingList.vue
@@ -5,25 +5,26 @@
                 <h4>Shopping Lists</h4>
                 <ul class="list-group pb-4">
                     <li v-for="(list,index) in shoppingLists" :key="list.id" class="list-group-item">
-                        <div class="col-9 float-left">
+                        <div class="col-8 float-left">
                             <a v-if="editListOffset!=index" href="javascript:void(0)" v-on:click="getList(list.id)">{{ list.name }}</a>
                             <div id="myId" ref="myId"></div>
                             <input type="text" 
-                                v-if="editListOffset==index" 
-                                @keyup.enter="updateList"
-                                v-on:blur="cancelEdit"
+                                v-show="editListOffset==index" 
+                                ref="foo"
+                                @keydown.enter="updateList"
+                                v-on:blur="cancelEdit(index)"
                                 class="form-control" 
                                 v-model="editList.name" 
-                                ref="listEdit">
+                                >
                         </div>
                         <div class="col-1 float-right text-right">
                             <a href="#" @click="deleteList(list.id)">
-                                <span class="fa fa-trash"></span> 
+                                <span class="fa fa-trash fa-lg"></span> 
                             </a>
                         </div>
-                        <div class="col-1 float-right text-right">
+                        <div class="col-2 float-right text-right">
                             <a href="#" @click="startEditing(index)">
-                                <span class="fa fa-pencil"></span> 
+                                <span class="fa fa-pencil fa-lg"></span> 
                             </a>
                         </div>                        
                     </li>
@@ -35,10 +36,17 @@
         </div>
         <div class="row" v-if="renderType=='detail'">
             <div class="col-12 pt-5">
-                <a href="javascript:void(0)" @click="back">Back to Lists</a>
+                <a href="javascript:void(0)" @click="back">&lt; Back to Lists</a>
             </div>
             <div class="col-12 pt-5">
-                <h4>{{ shoppingList.name }}</h4>
+                <div class="row">
+                    <div class="col-8">
+                        <h4>{{ shoppingList.name }}</h4>
+                    </div>
+                    <div class="col-4">
+                        {{ shoppingList.created_at | shortDate }}
+                    </div>
+                </div>
                 <ul class="list-group pb-4">
                     <li v-for="(item,index) in shoppingList.listitems" :key="index" class="list-group-item">
                         <div class="row">
@@ -47,7 +55,7 @@
                             </div>
                             <div class="col-2 text-right">
                                 <a href="#" @click="deleteItem(item.id)">
-                                    <span class="fa fa-trash"></span> 
+                                    <span class="fa fa-trash fa-lg red"></span> 
                                 </a>
                             </div>
                         </div>
@@ -72,12 +80,22 @@ export default {
             shoppingList: {},
             newList: null,
             newItem: null,
+            editListUneditedName: null,
             editList: {},
             editListOffset: -1,
         }
     },
     props: {
         appUrl: null,
+    },
+    filters: {
+        shortDate: function(value) {
+            if(!value) return ''
+
+            value = value.toString()
+
+            return moment(value).format('MMMM Do YYYY')
+        }
     },
     methods: {
         getLists() {
@@ -87,7 +105,7 @@ export default {
                     this.shoppingLists = response.data
                 })
                 .catch(error => {
-                    console.log(error)
+                    console.log('lists',error)
                 })
         },
         getList(id) {
@@ -100,7 +118,7 @@ export default {
                     this.renderType = 'detail'
                 })
                 .catch(error => {
-                    console.log(error)
+                    console.log('hello',error)
                 })
         },
         back() {
@@ -118,17 +136,25 @@ export default {
                     console.log(error)
                 })    
 
-            this.newList = null
+            this.newList = null 
         },
         startEditing(index) {
             
-            console.log(this.$refs)
             // this.$refs.listEdit.$el.focus()
             this.editListOffset = index
             this.editList = this.shoppingLists[index]
+            this.editListUneditedName = this.editList.name
+
+            console.log(this.$refs)
+            console.log(this.$refs.foo[index])
+
+            this.$nextTick(() => {
+                this.$refs.foo[index].focus()
+            })            
         },
         cancelEdit() {
             this.editListOffset = -1
+            this.editList.name = this.editListUneditedName
         },
         updateList() {
             let editListName = this.editList.name.charAt(0).toUpperCase()+this.editList.name.slice(1)
@@ -156,22 +182,26 @@ export default {
             this.newItem = null
         },
         deleteList(listId) {
-            axios.delete('/api/shoppinglists/'+listId)
-                .then(response => {
-                    this.getLists()
-                })                        
-                .catch(error => {
-                    console.log(error)
-                })
+            if(confirm('Delete List?')) {
+                axios.delete('/api/shoppinglists/'+listId)
+                    .then(response => {
+                        this.getLists()
+                    })                        
+                    .catch(error => {
+                        console.log(error)
+                    })
+            }
         },
         deleteItem(itemId) {
-            axios.delete('/api/shoppinglists/'+this.shoppingList.id+'/listitems/'+itemId)
-                .then(response => {
-                    this.getList(this.shoppingList.id)
-                })                        
-                .catch(error => {
-                    console.log(error)
-                })
+            if(confirm('Delete item?')) {
+                axios.delete('/api/shoppinglists/'+this.shoppingList.id+'/listitems/'+itemId)
+                    .then(response => {
+                        this.getList(this.shoppingList.id)
+                    })                        
+                    .catch(error => {
+                        console.log(error)
+                    })
+            }
         }
     },
     created() {
@@ -183,4 +213,18 @@ export default {
 
 }
 </script>
+
+<style>
+ul.list-group li a, ul.list-group li h6{
+    font-size: 1.3rem !important;
+    color: rgb(95, 95, 95);
+    
+}
+.fa-trash {
+    color: lightcoral;
+}
+.fa-pencil {
+    color: lightseagreen;
+}
+</style>
